+++ src/all.ts
@@ -9,6 +9,7 @@ export { CpsIconButton } from './components/icon-button.js';
 export { CpsInclude } from './components/include.js';
 export { CpsInput } from './components/input.js';
 export { CpsLabel } from './components/label.js';
+export { CpsNotification } from './components/notification.js';
 export { CpsPopover } from './components/popover.js';
 export { CpsRadio } from './components/radio.js';
 export { CpsRadioGroup } from './components/radio-group.js';