+++ src/components/notification/notification.styles.ts
@@ -1,0 +1,97 @@
+import { css } from 'lit';
+import componentStyles from '../../styles/component.styles.js';
+
+export default css`
+  ${componentStyles}
+
+  :host {
+    display: contents;
+
+    /* For better DX, we'll reset the margin here so the base part can inherit it */
+    margin: 0;
+  }
+
+  .notification {
+    display: flex;
+    position: relative;
+    align-items: stretch;
+    margin: inherit;
+    border: solid 1px var(--cps-stroke-card-primary);
+    border-radius: var(--cps-border-radius-medium);
+    background-color: var(--cps-fill-system-subtle-neutral);
+    color: var(--cps-foreground-primary);
+    font: var(--cps-text-body);
+  }
+
+  .notification:not(.notification--has-icon) .notification__icon,
+  .notification:not(.notification--closable) .notification__close-button {
+    display: none;
+  }
+
+  .notification__icon {
+    display: flex;
+    flex: 0 0 auto;
+    align-items: center;
+    justify-content: center;
+    margin: var(--cps-spacing-4) 0 0 var(--cps-spacing-3);
+    border-radius: var(--cps-border-radius-full);
+    width: var(--cps-spacing-4);
+    height: var(--cps-spacing-4);
+    color: rgb(var(--cps-color-neutral-0));
+    font-size: var(--cps-font-size-xs);
+  }
+
+  .notification--neutral {
+    background: var(--cps-fill-system-subtle-neutral);
+  }
+
+  .notification--neutral .notification__icon {
+    background: var(--cps-fill-system-neutral);
+  }
+
+  .notification--informative {
+    background: var(--cps-fill-system-subtle-informative);
+  }
+
+  .notification--informative .notification__icon {
+    background: var(--cps-fill-system-informative);
+  }
+
+  .notification--warning {
+    background: var(--cps-fill-system-subtle-warning);
+  }
+
+  .notification--warning .notification__icon {
+    background: var(--cps-fill-system-warning);
+  }
+
+  .notification--critical {
+    background: var(--cps-fill-system-subtle-critical);
+  }
+
+  .notification--critical .notification__icon {
+    background: var(--cps-fill-system-critical);
+  }
+
+  .notification--success {
+    background: var(--cps-fill-system-subtle-success);
+  }
+
+  .notification--success .notification__icon {
+    background: var(--cps-fill-system-success);
+  }
+
+  .notification__message {
+    display: block;
+    flex: 1 1 auto;
+    padding: var(--cps-spacing-4) var(--cps-spacing-3);
+    overflow: hidden;
+  }
+
+  .notification__close-button {
+    display: flex;
+    flex: 0 0 auto;
+    align-items: start;
+    margin: var(--cps-spacing-4) var(--cps-spacing-3) 0 0;
+  }
+`;